Update root page
